+++ test/client.test.js
@@ -1,11 +1,18 @@
 import Client from "../src/client";
-import testServer, { results } from "./test-server";
+import withTestServer, { results } from "./test-server";
 
-test("fetches the expected results", async () => {
-  testServer.listen(8080);
+test("sets sane defaults", () => {
+  const client = new Client();
 
-  const response = await new Client("token").search("Happy");
-  expect(response).toEqual(results);
-
-  testServer.close();
+  expect(client.base).toContain("api.tenor.com");
+  expect(typeof client.token).toEqual("string");
 });
+
+test("fetches the expected results", () => (
+  withTestServer(8083, async () => {
+    const client = new Client({ base: "http://localhost:8083", token: "token" });
+    const response = await client.search("Happy");
+
+    expect(response).toEqual(results);
+  })
+));