+++ test/test-server.js
@@ -8,12 +8,21 @@ export const results = [
   { id: 5, media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }] }
 ];
 
-const server = http.createServer();
+const withTestServer = async (port, callback) => {
+  const server = http.createServer();
+  server.requests = 0;
 
-server.on("request", (request, response) => {
-  response.writeHead(200, { "Content-Type": "application/json" });
-  response.write(JSON.stringify({ results }));
-  response.end();
-});
+  server.on("request", (request, response) => {
+    server.requests += 1;
+    response.writeHead(200, { "Content-Type": "application/json" });
+    response.write(JSON.stringify({ results }));
+    response.end();
+  });
 
-export default server;
+  server.listen(port);
+  await callback(server);
+
+  server.close();
+};
+
+export default withTestServer;